+++ .storybook/custom/components/DocsPage.js
@@ -1,9 +1,9 @@
+import '../custom.scss';
 import Community from './Community';
 import Footer from './Footer';
 import Header from './Header';
 import Import from './Import';
-import tocbot from 'tocbot';
-import React, { useEffect } from 'react';
+import React from 'react';
 import {
     Title,
     Subtitle,
@@ -14,9 +14,6 @@ import {
   } from '@storybook/addon-docs/blocks';
 
 const DocsPage = () => {
-    // useEffect(() => {
-    //     tocbot.init(Contents.tocOptions);
-    // }, []);
 
     return (
         <>
