+++ .storybook/custom/custom.scss
@@ -1,0 +1,189 @@
+body {
+    font-family: '72', 'Open Sans', sans-serif;
+}
+
+/* override storybook max width */
+.sbdocs.sbdocs-content {
+    max-width: 2000px;
+    padding-left: 20px;
+}
+
+/* override storybook section titles for color contrast */
+.simplebar-content .sidebar-subheading {
+    color: #404040;
+}
+
+/* header */
+.fr-header {
+    box-sizing: border-box;
+    height: 40px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    color: #404040;
+    background-color: #fff;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    border-bottom: 2px solid #d7dbde;
+    z-index: 10;
+
+    &--left {
+        display: flex;
+        align-items: center;
+        justify-self: start;
+        justify-content: space-between;
+    }
+
+    &--right {
+        justify-self: end;
+        display: flex;
+        justify-content: space-between;
+    }
+
+    span {
+        margin-left: 20px;
+    }
+
+    a {
+        padding: 10px;
+        color: #404040;
+        text-decoration: none;
+    }
+
+    a + a {
+        margin-top: 5px;
+    }
+
+    svg {
+        width: 1.5rem;
+        height: 1.5rem;
+    }
+}
+
+/* footer */
+.fr-footer {
+    padding: 16px 0;
+    font-size: 14px;
+    line-height: 1.33;
+    display: grid;
+    grid-template-columns: auto auto;
+    align-items: center;
+    border-top: 2px solid #d7dbde;
+    color: #404040;
+
+    &__logo {
+        margin-right: 16px;
+        min-width: 120px;
+        display: flex;
+        align-items: center;
+    }
+
+    &__slogan {
+        white-space: nowrap;
+        margin-right: 5px;
+        font-weight: bold;
+    }
+
+    &__copyright {
+        margin-left: 24px;
+        width: 100%;
+    }
+
+    &--left {
+        display: flex;
+        align-items: center;
+        justify-self: start;
+        justify-content: space-between;
+    }
+
+    &--right {
+        justify-self: end;
+        display: flex;
+        justify-content: space-between;
+
+        a {
+            margin-left: 24px;
+            text-decoration: none;
+            color: #404040;
+
+            &:hover {
+                color: #0e7290;
+            }
+        }
+    }
+}
+
+/* community social links */
+.fr-community {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    border-top: 2px solid #d7dbde;
+    color: #404040;
+    font-size: 14px;
+
+    section {
+        width: 50%;
+        margin: 10px;
+    }
+
+    ul {
+        list-style-type: none;
+        padding-left: 0;
+    }
+
+    li {
+        margin: 0 0 20px 0;
+    }
+
+    h4 {
+        font-weight: normal;
+        font-size: 16px;
+    }
+
+    a {
+        color: #0e7290;
+        text-decoration: none;
+
+        &:hover {
+            color: #404040;
+        }
+    }
+
+    img {
+        height: 18px;
+        vertical-align: middle;
+        margin-right: 10px;
+    }
+}
+
+/* story example layout containers */
+.fr-container {
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -ms-flex-wrap: wrap;
+    flex-wrap: wrap;
+    -ms-flex-pack: distribute;
+    justify-content: space-around;
+    width: 100%;
+    padding: 10px 10px;
+}
+
+.break {
+    flex-basis: 100%;
+    height: 0;
+    border-bottom: 10px solid transparent;
+}
+
+.demo-icon-wrapper {
+    padding: 10px;
+    float: left;
+    width: 16%;
+    text-align: center;
+    height: 120px;
+}
