+++ package.json
@@ -21,7 +21,7 @@
     "lint:fix": "npm run lint:code:fix && npm run lint:style:fix",
     "lint:code": "eslint . --ext .js,.jsx",
     "lint:code:fix": "npm run lint:code -- --fix",
-    "lint:style": "stylelint \".storybook/custom/*.css\"",
+    "lint:style": "stylelint \".storybook/custom/*.scss\"",
     "lint:style:fix": "npm run lint:style -- --fix",
     "lint:pre-commit": "printf \"running pre-commit lint...\"  && npm run lint && printf \"done!\n\"",
     "release:create": "create-release",
@@ -92,6 +92,7 @@
     "core-js": "^3.1.4",
     "coveralls": "^3.0.2",
     "cross-env": "^7.0.2",
+    "css-loader": "^3.5.3",
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.7.0",
     "eslint": "^6.8.0",
@@ -109,6 +110,7 @@
     "github-assistant": "^0.3.1-rc.0",
     "jest": "24.9.0",
     "jest-puppeteer": "^4.4.0",
+    "node-sass": "^4.14.1",
     "pre-commit": "^1.2.2",
     "prettier": "^1.19.1",
     "puppeteer": "^3.0.0",
@@ -118,10 +120,12 @@
     "react-syntax-highlighter": "^12.2.1",
     "react-test-renderer": "^16.6.3",
     "regenerator-runtime": "^0.13.2",
+    "sass-loader": "^8.0.2",
     "size-limit": "^4.0.1",
     "standard-version": "^7.1.0",
     "start-server-and-test": "^1.10.6",
     "storybook-addon-i18n": "^5.1.13",
+    "style-loader": "^1.2.1",
     "stylelint": "^13.3.3",
     "stylelint-config-prettier": "^8.0.0",
     "stylelint-prettier": "^1.1.2",
